refactored CustomSelect
